// frontend/src/api/generated/models/CreateTask.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Inundated
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateTask
 */
export interface CreateTask {
    /**
     * 
     * @type {string}
     * @memberof CreateTask
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof CreateTask
     */
    parentId?: string;
    /**
     * 
     * @type {Set<string>}
     * @memberof CreateTask
     */
    tagIds?: Set<string>;
    /**
     * 
     * @type {Date}
     * @memberof CreateTask
     */
    dueDate?: Date;
    /**
     * 
     * @type {number}
     * @memberof CreateTask
     */
    estimateHours?: number;
}

/**
 * Check if a given object implements the CreateTask interface.
 */
export function instanceOfCreateTask(value: object): value is CreateTask {
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function CreateTaskFromJSON(json: any): CreateTask {
    return CreateTaskFromJSONTyped(json, false);
}

export function CreateTaskFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateTask {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'parentId': json['parentId'] == null ? undefined : json['parentId'],
        'tagIds': json['tagIds'] == null ? undefined : new Set(json['tagIds']),
        'dueDate': json['dueDate'] == null ? undefined : (new Date(json['dueDate'])),
        'estimateHours': json['estimateHours'] == null ? undefined : json['estimateHours'],
    };
}

export function CreateTaskToJSON(json: any): CreateTask {
    return CreateTaskToJSONTyped(json, false);
}

export function CreateTaskToJSONTyped(value?: CreateTask | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'parentId': value['parentId'],
        'tagIds': value['tagIds'] == null ? undefined : Array.from(value['tagIds'] as Set<any>),
        'dueDate': value['dueDate'] == null ? value['dueDate'] : value['dueDate'].toISOString().substring(0,10),
        'estimateHours': value['estimateHours'],
    };
}