// frontend/src/api/generated/models/UpdateTask.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Inundated
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { CloseReason } from './CloseReason';
import {
    CloseReasonFromJSON,
    CloseReasonFromJSONTyped,
    CloseReasonToJSON,
    CloseReasonToJSONTyped,
} from './CloseReason';

/**
 * 
 * @export
 * @interface UpdateTask
 */
export interface UpdateTask {
    /**
     * 
     * @type {string}
     * @memberof UpdateTask
     */
    name?: string;
    /**
     * 
     * @type {Set<string>}
     * @memberof UpdateTask
     */
    tagIds?: Set<string>;
    /**
     * 
     * @type {Date}
     * @memberof UpdateTask
     */
    dueDate?: Date;
    /**
     * Removes the due date. Can't be combined with dueDate.
     * @type {boolean}
     * @memberof UpdateTask
     */
    clearDueDate?: boolean;
    /**
     * 
     * @type {number}
     * @memberof UpdateTask
     */
    estimateHours?: number;
    /**
     * Removes the estimate. Can't be combined with estimateHours.
     * @type {boolean}
     * @memberof UpdateTask
     */
    clearEstimate?: boolean;
    /**
     * Closing a task also closes its open subtasks with the same reason. Reopening a task also reopens its closed ancestors.
     * 
     * @type {boolean}
     * @memberof UpdateTask
     */
    closed?: boolean;
    /**
     * 
     * @type {CloseReason}
     * @memberof UpdateTask
     */
    closeReason?: CloseReason;
}



/**
 * Check if a given object implements the UpdateTask interface.
 */
export function instanceOfUpdateTask(value: object): value is UpdateTask {
    return true;
}

export function UpdateTaskFromJSON(json: any): UpdateTask {
    return UpdateTaskFromJSONTyped(json, false);
}

export function UpdateTaskFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateTask {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'tagIds': json['tagIds'] == null ? undefined : new Set(json['tagIds']),
        'dueDate': json['dueDate'] == null ? undefined : (new Date(json['dueDate'])),
        'clearDueDate': json['clearDueDate'] == null ? undefined : json['clearDueDate'],
        'estimateHours': json['estimateHours'] == null ? undefined : json['estimateHours'],
        'clearEstimate': json['clearEstimate'] == null ? undefined : json['clearEstimate'],
        'closed': json['closed'] == null ? undefined : json['closed'],
        'closeReason': json['closeReason'] == null ? undefined : CloseReasonFromJSON(json['closeReason']),
    };
}

export function UpdateTaskToJSON(json: any): UpdateTask {
    return UpdateTaskToJSONTyped(json, false);
}

export function UpdateTaskToJSONTyped(value?: UpdateTask | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'tagIds': value['tagIds'] == null ? undefined : Array.from(value['tagIds'] as Set<any>),
        'dueDate': value['dueDate'] == null ? value['dueDate'] : value['dueDate'].toISOString().substring(0,10),
        'clearDueDate': value['clearDueDate'],
        'estimateHours': value['estimateHours'],
        'clearEstimate': value['clearEstimate'],
        'closed': value['closed'],
        'closeReason': CloseReasonToJSON(value['closeReason']),
    };
}